Implement pet management functions: create, delete, get, and update with validation; add authorization handling and exceptions

+++ functions/src/features/user/use_cases/delete/index.ts
@@ -2,20 +2,33 @@ import { onRequest } from "firebase-functions/https";
 import { DeleteUserModel } from "./models/DeleteUserModel";
 import { ExceptionsHandler } from "../../../../shared/handlers/ExceptionsHandler";
 import { firestore } from "../../../../shared/firestore/init";
+import { validateAuthToken } from "../../../../shared/auth/validateAuthToken";
+import { FirestoreCollections } from "../../../../shared/firestore/collections";
 
 export const deleteUser = onRequest(async (req, res) => {
     try {
         const body = new Map(Object.entries(req.body));
         const model = new DeleteUserModel(body);
 
-        const doc = firestore.collection('users').doc(model.id);
+        const oauthToken = req.headers.authorization;
+        const userIdFromToken = await validateAuthToken(oauthToken?.replace('Bearer ', ''));
+        
+        if(model.id !== userIdFromToken) {
+            res.status(403).json({ message: 'You can only delete your own user account' });
+            return;
+        }
 
-        doc.get().then((doc) => {
-            if (!doc.exists) {
-                res.status(404).json({ message: 'User not found' });
-                return;
-            }
-        });
+        const doc = firestore.collection(FirestoreCollections.Users).doc(userIdFromToken);
+        if(!(await doc.get()).exists) {
+            res.status(404).json({ message: 'User not found' });
+            return;
+        }
+
+        const data = (await doc.get()).data();
+        if(data?.email !== model.email || data?.googleId !== model.googleId || data?.id !== model.id) {
+            res.status(400).json({ message: 'User data does not match' });
+            return;
+        }
 
         doc.delete()
             .then(() => {
